rename app -> script_runner
this is going to be a library so we should give it a better name

we don't want people to have to `import app` as that is too ambiguous
