+++ script_runner/app.py
@@ -6,8 +6,8 @@ from typing import Any, Callable
 import requests
 from flask import Flask, Response, jsonify, request, send_from_directory
 
-from app.auth import UnauthorizedUser
-from app.utils import CombinedConfig, MainConfig, RegionConfig, load_config
+from script_runner.auth import UnauthorizedUser
+from script_runner.utils import CombinedConfig, MainConfig, RegionConfig, load_config
 
 app = Flask(__name__)
