+++ tests/test_utils.py
@@ -1,5 +1,5 @@
-from app.auth import GoogleAuth, NoAuth
-from app.utils import CommonFields, FunctionParameter, load_group
+from script_runner.auth import GoogleAuth, NoAuth
+from script_runner.utils import CommonFields, FunctionParameter, load_group
 
 
 def test_common_fields() -> None:
